--- src/main/java/com/mec/libapi/entity/User.java
@@ -1,23 +1,0 @@
-package com.mec.libapi.entity;
-
-import lombok.*;
-
-import javax.persistence.*;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Builder
-@Inheritance
-@Table(name = "users")
-@Entity
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String firstName;
-    private String secondName;
-    @Column(unique = true)
-    private String email;
-    private String password;
-}
