+++ src/main/java/com/mec/libapi/entity/bookModule/Availability.java
