+++ src/main/java/com/mec/libapi/entity/bookModule/Detail.java
