+++ src/main/java/com/mec/libapi/entity/bookModule/NCase.java
