+++ src/main/java/com/mec/libapi/entity/bookModule/Speciality.java
@@ -1,5 +1,6 @@
 package com.mec.libapi.entity;
 
+import com.mec.libapi.entity.bookmodule.Book;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
