+++ src/main/java/com/mec/libapi/entity/userModule/Professor.java
@@ -1,10 +1,10 @@
 package com.mec.libapi.entity;
 
+import com.mec.libapi.entity.userModule.CanBook;
 import lombok.*;
 
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
-import javax.persistence.Table;
 
 @NoArgsConstructor
 @AllArgsConstructor
