+++ src/main/java/com/mec/libapi/entity/userModule/Student.java
@@ -1,5 +1,6 @@
 package com.mec.libapi.entity;
 
+import com.mec.libapi.entity.userModule.CanBook;
 import lombok.*;
 
 import javax.persistence.*;
