+++ src/main/java/com/mec/libapi/repository/bookModule/NCaseRepository.java
@@ -1,12 +1,13 @@
 package com.mec.libapi.repository;
 
-import com.mec.libapi.entity.User;
+import com.mec.libapi.entity.bookModule.NCase;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmailAndPassword(String email, String password);
+public interface NCaseRepository extends JpaRepository<NCase, Long> {
+
+    Optional<NCase> findByCode(String code);
 }
