+++ src/main/java/com/mec/libapi/repository/bookModule/SpecialityRepository.java
@@ -1,9 +1,10 @@
 package com.mec.libapi.repository;
 
-import com.mec.libapi.entity.CanBook;
+import com.mec.libapi.entity.bookModule.Speciality;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CanBookRepository extends JpaRepository<CanBook, Long> {
+public interface SpecialityRepository extends JpaRepository<Speciality, Long> {
+
 }
