+++ src/main/java/com/mec/libapi/repository/userModule/CanBookRepository.java
@@ -1,9 +1,9 @@
 package com.mec.libapi.repository;
 
-import com.mec.libapi.entity.Supervisor;
+import com.mec.libapi.entity.userModule.CanBook;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SupervisorRepository extends JpaRepository<Supervisor, Long> {
+public interface CanBookRepository extends JpaRepository<CanBook, Long> {
 }
