+++ src/main/java/com/mec/libapi/repository/userModule/FineRepository.java
@@ -1,6 +1,6 @@
 package com.mec.libapi.repository;
 
-import com.mec.libapi.entity.Fine;
+import com.mec.libapi.entity.userModule.Fine;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
