+++ src/main/java/com/mec/libapi/repository/userModule/ProfessorRepository.java
@@ -1,9 +1,9 @@
 package com.mec.libapi.repository;
 
-import com.mec.libapi.entity.Student;
+import com.mec.libapi.entity.userModule.Professor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Long> {
+public interface ProfessorRepository extends JpaRepository<Professor, Long> {
 }
