+++ src/main/java/com/mec/libapi/repository/userModule/StudentRepository.java
@@ -1,9 +1,9 @@
 package com.mec.libapi.repository;
 
-import com.mec.libapi.entity.Professor;
+import com.mec.libapi.entity.userModule.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfessorRepository extends JpaRepository<Professor, Long> {
+public interface StudentRepository extends JpaRepository<Student, Long> {
 }
