+++ src/main/java/com/mec/libapi/repository/userModule/SupervisorRepository.java
@@ -1,10 +1,9 @@
 package com.mec.libapi.repository;
 
-import com.mec.libapi.entity.Speciality;
+import com.mec.libapi.entity.userModule.Supervisor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpecialityRepository extends JpaRepository<Speciality, Long> {
-
+public interface SupervisorRepository extends JpaRepository<Supervisor, Long> {
 }
