+++ src/main/java/com/mec/libapi/repository/userModule/UserRepository.java
@@ -1,13 +1,12 @@
 package com.mec.libapi.repository;
 
-import com.mec.libapi.entity.NCase;
+import com.mec.libapi.entity.userModule.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface NCaseRepository extends JpaRepository<NCase, Long> {
-
-    Optional<NCase> findByCode(String code);
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmailAndPassword(String email, String password);
 }
